Get rid of JavaMethodType
Its main purpose was to compute MethodType#signature but this is no
longer the case as of the previous commit which delegates this
responsability to SingleDenotation#signature, and it turns out that in
all places where we called `isJavaMethod` we can rely on contextual
informations to obtain the same information, so we can completely get
rid of it which is a nice simplification.

Also get rid of ElimRepeated#transformTypeOfTree which turned out to be
a no-op: the denotation transformer already transforms method types and
there's nothing to do on top of that.