Denotation#signature: Compute the Java-ness of the prefix more precisely
Instead of relying on `prefix.classSymbol.is(JavaDefined)`, we now
explicitly check if every part of the prefix is Java-defined.

Being precise like this doesn't seem necessary in practice (we support
having multiple denotations with the same signature since the
introduction of SELECTin in tasty), but it's important that we stay
consistent across versions for tasty compatibility so it's better to be
explicit than rely on the implementation details of `classSymbol`.